Moved CorePlugin to core_plugin.rs
